+++ energy_demand/profiles/generic_shapes.py
@@ -28,15 +28,15 @@ def flat_shape(nr_of_days):
     shape_peak_dh = np.full((24), 1/24)
 
     # linear shape_non_peak_y_dh
-    shape_non_peak_y_dh = np.zeros((nr_of_days, 24))
+    shape_non_peak_y_dh = np.zeros((nr_of_days, 24), dtype=float)
 
     # Flat shape, ever hour same amount
-    shape_non_peak_y_dh = np.full((nr_of_days, 24), (1.0/24))
+    shape_non_peak_y_dh = np.full((nr_of_days, 24), (1.0/24), dtype=float)
 
     # linear shape_non_peak_yd
     shape_non_peak_yd = np.ones((nr_of_days)) / nr_of_days
 
-    shape_non_peak_yh = np.full((nr_of_days, 24), 1/(nr_of_days * 24))
+    shape_non_peak_yh = np.full((nr_of_days, 24), 1/(nr_of_days * 24), dtype=float)
 
     return shape_peak_dh, shape_non_peak_y_dh, shape_peak_yd_factor, shape_non_peak_yd, shape_non_peak_yh
